Fix 64-bit pointer truncation in Live2D aligned allocator
The alignment mask ~(alignment - 1) was computed in 32-bit
(csmUint32), zeroing the upper 32 bits of 64-bit pointers.
Cast alignment to size_t before computing the mask.

+++ Main/src/Live2D/Live2DManager.cpp
@@ -29,7 +29,7 @@ namespace Live2D
 		if (!raw) return nullptr;
 
 		void** aligned = reinterpret_cast<void**>(
-			(reinterpret_cast<size_t>(raw) + offset) & ~(alignment - 1)
+			(reinterpret_cast<size_t>(raw) + offset) & ~(static_cast<size_t>(alignment) - 1)
 		);
 		aligned[-1] = raw;
 		return aligned;
